Create a docker mod to add rbenv to code-server
This mod adds rbenv to code-server, to be installed/updated during container start.

rbenv is a version manager tool for the Ruby programming language on Unix-like systems. It is useful for switching between multiple Ruby versions on the same machine and for ensuring that each project you are working on always runs on the correct Ruby version.

This mod includes adding shell completions for rbenv in bash and zsh.